extracted access key strategy
